bpf: add stand-alone load-balancer ipip health check datapath
Add the possibility to health probe backends through ipip encap so they
reach endpoints in the same way as regular datapath traffic.

There are various options on how this could be achieved. One possibility
we considered was to use SRIOV where we'd spawn a VF and inject traffic
into the regular XDP datapath at the extrnal facing device by rewriting
the dst MAC to the one of the PF at VF tc egress time. This traffic then
reaches the XDP prog and can be pushed out the same devices for leaving
the node. This option has one downside though, that is, it's limited to
bare metal deployment scenarios (it's not possible to spawn a VF from VF
when in off-prem / cloud) and the HW needs to support SRIOV. Performance
wise, the use of XDP_TX has no downgrade under SRIOV, but XDP_REDIRECT
has. In Cilium case we only use XDP_TX today, but in future we potentially
could also extend for multi-dev XDP_REDIRECT case.

This led us to a more flexible tc BPF based approach which does not have
the latter limitations. We do rely on ipip / ip6ip6 tunnel driver in
collect metadata mode though, meaning, this adds cilium_ipip4, cilium_ipip6
and cilium_sit devices. From kernel side, ipip driver adds the three netdevs
as a pair. One more design consideration is that the health check probes
are decoupled from the BPF service map, meaning, the service map is not
consulted on purpose in the datapath - this has the advantage that backends
can be independently checked from the service map without having to have a
'staging phase' for the service map to push them live after initial probe.
Additionally, the service map doesn't need a more complex synchronization
scheme with the health check logic to prevent backend position reordering
during a health probe run (for telling which of the backend failed when
they would be picked based on socket mark, for example).

That said, if the health checker residing on the LB node would initiate a
connection to svcIP:port, we still need a way to pass-in a backend IP that
we intend to test for availability, so that the ipip outer header can then
be populated as [L2][lbIP -> backendIP][lbIP -> svcIP][L4]. Typical client
applications allocate a socket and then call connect(2) or sendmsg(2) for
the probe. The connect(2) call will auto-bind the socket in the kernel
for obtaining a source port. This could be seen as equivalent of calling
bind(0.0.0.0:0) on the client before calling connect(2) or sendmsg(2). We
can piggy-back on this, mark a socket for getting special treatment for
the health check datapath and then do i) bind(backendIP:port) with ii)
subsequent connect(serviceIP:port). The bind()ing to a backend approach
allows for BPF at the pre-bind hook to record a IPv4 or IPv6 backend address
to be recorded in an LRU map based on the socket cookie where later on a
tc egress BPF program can retrieve that information from the skb (again,
via socket cookie) to populate the collect metadata tunnel key. The pre-bind
hook in the kernel allows to override the struct sockaddr, which allows to
clear the IP/port in order to let the kernel perform auto-bind. The socket
itself is locked during this time. This allows to define a simple API for
backend health probes.

Arguments used for testing:

  # ./daemon/cilium-agent --enable-ipv4=true --enable-ipv6=true \
       --datapath-mode=lb-only --bpf-lb-algorithm=maglev        \
       --bpf-lb-maglev-table-size=2039 --bpf-lb-mode=dsr        \
       --bpf-lb-acceleration=native --devices=enp2s0np0         \
       --bpf-lb-dsr-dispatch=ipip --disable-envoy-version-check=true

Example health probe client in C has been added outside of the main repo:

  https://github.com/cilium/health-examples

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>